% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMetaObject.R
\name{AddMetaObject}
\alias{AddMetaObject}
\title{adding cell assignments to seurat object}
\usage{
AddMetaObject(seurat_obj, cell_class_df)
}
\arguments{
\item{seurat_obj}{a prepossessed Seurat object storing the scRNAseq data}

\item{cell_class_df}{a data frame of cell barcodes and cell type assignments, ideally obtained as the output from the GetCellAssignments() function.}
}
\value{
a new Seurat object with the updated metadata containing predicted cell labels in the "celltype_ulm" column
}
\description{
'AddMtaObject()' a function to add the predicted cell labels to the metadata of scRNAseq
}
\examples{
data(int_multData)
data(int_signature)
my_scores <- GetCellScores(seurat_obj = int_multData[,1:1000], signatures = int_signature, assay = 'RNA', slot = 'data')
my_ass <- GetCellAssignments(score_data = my_scores)
new_obj <- AddMetaObject(seurat_obj = int_multData[,1:1000], cell_class_df = my_ass)
head(new_obj$celltype_ulm, 20)

}
