% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMultiplet.R
\name{FilterMultiplet}
\alias{FilterMultiplet}
\title{filtering multiplets}
\usage{
FilterMultiplet(seurat_obj, minCells = 2, minFreq = 10)
}
\arguments{
\item{seurat_obj}{a Seurat object with the metadata containing predicted cell labels in the "celltype_ulm" column and the number of cells in the "count_ulm" column. Ideally the output from the AddMetObject() function or the multiplet Seurat object from the GetMultiplet() function}

\item{minCells}{a numeric value specifying the minimum number of cells. Default is 2 to include doublets and/or higher order multiplets}

\item{minFreq}{a numeric value specifying the minimum frequency of a multiplet type for it to be retained. Default is 10.}
}
\value{
a filtered list containing a Seurat object of multiplets and a dataframe of multiplet distribution summary.
}
\description{
'FilterMultiplet()' a function to filter multiplets of a predefined frequency
}
\examples{

data(int_multData)
data(int_signature)
my_scores <- GetCellScores(seurat_obj = int_multData[,1:1000], signatures = int_signature, assay = 'RNA', slot = 'data')
my_ass <- GetCellAssignments(score_data = my_scores)
new_obj <- AddMetaObject(seurat_obj = int_multData[,1:1000], cell_class_df = my_ass)
my_mult <- GetMultiplet(seurat_obj = new_obj)
my_mult_filt <- FilterMultiplet(seurat_obj = new_obj)
my_mult_filt

}
