% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellAssignments.R
\name{GetCellAssignments}
\alias{GetCellAssignments}
\title{getting final cell assignments}
\usage{
GetCellAssignments(score_data, p_val = 0.05, cut_off = 1)
}
\arguments{
\item{score_data}{a data frame of cell barcodes and gene signature scores, ideally obtained as the output from the GetCellScores() function.}

\item{p_val}{a numeric value specifying the p value cut-off to filter significant signature scores (default: 0.05)}

\item{cut_off}{a numeric value specifying the cut-off for signature scores (default: 1)}
}
\value{
a data frame of cell barcodes and cell type assignments. A barcode may be assigned a single or multi cell type assignment depending on signature enrichment scores.
}
\description{
'GetCellAssignments()' a function that assigns cell type labels to each barcode based on the signature scores
}
\examples{
data(int_singData)
data(int_signature)
my_scores <- GetCellScores(seurat_obj = int_singData[,1:1000], signatures = int_signature, assay = 'RNA', slot = 'data')
my_ass <- GetCellAssignments(score_data = my_scores)
head(my_ass)

}
