% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellScores.R
\name{GetCellScores}
\alias{GetCellScores}
\title{scoring cells for gene signatures}
\usage{
GetCellScores(
  seurat_obj,
  signatures,
  assay = "RNA",
  slot = NULL,
  layer = "data"
)
}
\arguments{
\item{seurat_obj}{a preprocessed Seurat object storing the scRNAseq data}

\item{signatures}{a data frame of cell type signatures with cell types in the source column, genes in the target column, and weights in the mor column. Ideally the output from the GetSignature() function.}

\item{assay}{a character specifying assay in the Seurat object containing count matrix.}

\item{layer}{a character specifying the layer to draw counts from (Seurat v3/v4). This can be "counts" for raw counts, "data" for normalized counts, or "scaled" for scaled counts. Default is NULL.}
}
\value{
a data frame of cell barcodes and gene signature scores
}
\description{
'GetCellScores()' scores each cell in the scRNAseq data for cell type-specific gene signatures
}
\examples{
data(int_singData)
data(int_signature)
my_scores <- GetCellScores(seurat_obj = int_singData[,1:1000],
                           signatures = int_signature,
                           assay = 'RNA',
                           layer = 'data')
head(my_scores)

}
