% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMultiplet.R
\name{GetMultiplet}
\alias{GetMultiplet}
\title{obtaining multiplet data}
\usage{
GetMultiplet(seurat_obj, minCells = 2)
}
\arguments{
\item{seurat_obj}{a Seurat object with the metadata containing predicted cell labels in the "celltype_ulm" column and the number of cells in the "count_ulm" column. Ideally the output from the AddMetObject() function}

\item{minCells}{a numeric value specifying the minimum number of cells. Default is 2 to include doublets and/or higher order multiplets}
}
\value{
a list containing a Seurat object of multiplets and multiplet distribution summary
}
\description{
'GetMultiplet()' a function to extract predicted multiplets
}
\examples{

data(int_multData)
data(int_signature)
my_scores <- GetCellScores(seurat_obj = int_multData[,1:1000], signatures = int_signature, assay = 'RNA', layer = 'data')
my_ass <- GetCellAssignments(score_data = my_scores)
new_obj <- AddMetaObject(seurat_obj = int_multData[,1:1000], cell_class_df = my_ass)
my_mult <- GetMultiplet(seurat_obj = new_obj)
my_mult



}
