% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNodeDF.R
\name{GetNodeDF}
\alias{GetNodeDF}
\title{getting nodes and edges}
\usage{
GetNodeDF(mat)
}
\arguments{
\item{mat}{a data frame of multiplets types and their frequencies, ideally the summary data frame obtained from the FilterMultiplet() or GetMultiplet() function.}
}
\value{
a data frame of cell-cell edges and nodes and the corresponding frequencies
}
\description{
'GetNodeDF()' a function to generate nodes and edges from multiplet data
}
\examples{
mat <- data.frame('multilpetType' = c(paste('A', 'B', 'C', sep = '_'),
                                      paste('A', 'B', sep = '_'),
                                      paste('B', 'C', 'A', 'E', sep = '_'),
                                      paste('D', 'C', 'E', sep = '_')),
                  frequency = rep(50, 4))

mat
GetNodeDF(mat)

}
