% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSignature.R
\name{GetSignature}
\alias{GetSignature}
\title{generating gene signature}
\usage{
GetSignature(seurat_obj, ident_col = NULL, n = 100, p_val = 0.05)
}
\arguments{
\item{seurat_obj}{a prepossessed Seurat object storing the scRNAseq data}

\item{ident_col}{a column in the Seurat object metadata which is a character vector storing cell names or labels. If not specified, the default ident of the Seurat object will be used.}

\item{n}{a numeric value specifying the number of genes to be used for each cell signature. Default is 100 genes per cell type.}

\item{p_val}{a numeric value specifying the adjusted p-value cut-off}
}
\value{
a dataframe of cell type signatures.
}
\description{
'GetSignature()' generates cell type-specific gene signatures from scRNAseq data
}
\examples{
data(int_singData)
int_sig <- GetSignature(seurat_obj = int_singData[,1:1000], ident_col = int_singData$Cell_Type)
head(int_sig)

}
